motorSpeedPWM is an Int, not Double. Small addition to previously commited fix (667be6c)

+++ src/replicatorg/app/ui/controlpanel/ExtruderPanel.java
@@ -162,7 +162,7 @@ public class ExtruderPanel extends JPanel implements FocusListener, ActionListen
 				field.setName("motor-speed-pwm");
 				field.addFocusListener(this);
 				field.setActionCommand("handleTextfield");
-				field.setText(Double.toString(t.getMotorSpeedPWM()));
+				field.setText(Integer.toString(t.getMotorSpeedPWM()));
 				field.addActionListener(this);
 
 				add(label);
